@@ -1,14 +1,26 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import React, { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import "./index.css";
+import App from "./App";
+import * as serviceWorker from "./serviceWorker";
+import { Amplify } from "aws-amplify";
+import awsconfig from "./aws-exports";
+import { Authenticator } from "@aws-amplify/ui-react";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+Amplify.Logger.LOG_LEVEL = "DEBUG";
+
+Amplify.configure(awsconfig);
+
+const container = document.getElementById("app");
+
+const root = createRoot(container);
+
+root.render(
+  <StrictMode>
+    <Authenticator.Provider>
+      <App />
+    </Authenticator.Provider>
+  </StrictMode>
 );
 
 // If you want your app to work offline and load faster, you can change
